Modified the util method

+++ components/tenant-mgt/org.wso2.carbon.tenant.mgt/src/main/java/org/wso2/carbon/tenant/mgt/util/TenantMgtUtil.java
@@ -322,41 +322,41 @@ public class TenantMgtUtil {
 
     /**
      * @param tenantDomain domain name of the tenant.
-     * @param tenantManager tenant manager used for tenant deletion .
-     * @param tenantId tenant id.
-     * @param serverConfigurationService serverConfigurationService.
      * @throws Exception if there is an exception during the tenant deletion.
      */
-    public static void deleteTenant(String tenantDomain, TenantManager tenantManager, int tenantId,
-                                    ServerConfigurationService serverConfigurationService) throws Exception {
+    public static void deleteTenant(String tenantDomain) throws Exception {
 
-    /*
-     * TODO: 2/7/19 We need to fix listeners to enable this by default
-     * Refer - https://github.com/wso2-support/carbon-multitenancy/issues/35
-     */
-        if (Boolean.parseBoolean(
-                serverConfigurationService.getFirstProperty("Tenant.ListenerInvocationPolicy.InvokeOnDelete"))) {
-            triggerPreTenantDelete(tenantId);
-        } else {
-            if (log.isDebugEnabled()) {
-                log.debug(
-                        "Tenant.ListenerInvocationPolicy.InvokeOnDelete flag is not set to true in carbon.xml. " +
-                                "Listener invocation ignored.");
+        TenantManager tenantManager = TenantMgtServiceComponent.getTenantManager();
+        if (tenantManager != null) {
+            int tenantId = tenantManager.getTenantId(tenantDomain);
+            ServerConfigurationService serverConfigurationService =
+                    TenantMgtServiceComponent.getServerConfigurationService();
+            /*
+             * TODO: 2/7/19 We need to fix listeners to enable this by default
+             */
+            if (Boolean.parseBoolean(
+                    serverConfigurationService.getFirstProperty("Tenant.ListenerInvocationPolicy.InvokeOnDelete"))) {
+                triggerPreTenantDelete(tenantId);
+            } else {
+                if (log.isDebugEnabled()) {
+                    log.debug("Tenant.ListenerInvocationPolicy.InvokeOnDelete flag is not set to true in " +
+                            "carbon.xml. Listener invocation ignored.");
+                }
             }
+
+            TenantMgtUtil.deleteWorkernodesTenant(tenantId);
+
+            if (TenantMgtServiceComponent.getBillingService() != null) {
+                TenantMgtServiceComponent.getBillingService().deleteBillingData(tenantId);
+            }
+
+            TenantMgtUtil.unloadTenantConfigurations(tenantDomain, tenantId);
+            TenantMgtUtil.deleteTenantRegistryData(tenantId);
+            TenantMgtUtil.deleteTenantDir(tenantId);
+            tenantManager.deleteTenant(tenantId);
+            log.info(String.format("Deleted tenant with domain: %s and tenant id: %d from the system.", tenantDomain,
+                    tenantId));
         }
-
-        TenantMgtUtil.deleteWorkernodesTenant(tenantId);
-
-        if (TenantMgtServiceComponent.getBillingService() != null) {
-            TenantMgtServiceComponent.getBillingService().deleteBillingData(tenantId);
-        }
-
-        TenantMgtUtil.unloadTenantConfigurations(tenantDomain, tenantId);
-        TenantMgtUtil.deleteTenantRegistryData(tenantId);
-        TenantMgtUtil.deleteTenantDir(tenantId);
-        tenantManager.deleteTenant(tenantId);
-        log.info(String.format("Deleted tenant with domain: %s and tenant id: %d from the system.", tenantDomain,
-                tenantId));
     }
 
     /**
